match coda-oss naming conventions (#688)
* std::complex<short> is no longer valid C++

* implement abs() for complex_short

* trying to fix abs() unittest

* trying to get GCC results to match

* still can't get right GCC result :-(

* trying to get abs() working with GCC

* reduce "visibility" of complex<short> #pragma

* complex_short -> complex

* types::complex<T> instead of types::complex_short

* Have the compiler pick between std::complex and details::complex

* zfloat, etc. aliases

* tweak getPrecision() for std::complex

* create our types::complex from std::complex

* Revert "tweak getPrecision() for std::complex"

This reverts commit 1c9396893067700e1d364a565f47101c4d6c74b5.

* use std::complex<int> for now until everything is switched over

* types::complex<T> is std::complex<T> for now

* str::toString() need to know about our new complex type

* try to prevent std::complex<short> from popogating

* test_toStringComplexFloat

* test_toStringComplexShort

* str::toString() for types::complex

* trying to fix compile warning

* trying to fix compile warning

* does specialization give a better type than std::conditional?

* operator==()

* types::complex<T> maybe isn't a great idea?

* be sure str::toType() works for our complex type

* comments/whitespace

* add zinteger and zreal

* test str::toString() and str::toType() for complex

* overload on std::complex<T> for existing code

* macros to determine whether zinteger should be std::complex or our own type

* use CODA_OSS_types_unique_zinteger macro

* update comments to match code

* use copy constructor to avoid compiler warning

* std::complex<short> can still be used, if desired.

* complex.h -> Complex.h

* zintegerT -> Complex

* overload assignment to match copy

* implicit conversion to std::complex ... <shrug>

* complex.h -> Complex.h
