+++ modules/c++/coda-oss-lite.vcxproj
@@ -256,7 +256,7 @@
     <ClInclude Include="tiff\include\tiff\KnownTags.h" />
     <ClInclude Include="tiff\include\tiff\TypeFactory.h" />
     <ClInclude Include="tiff\include\tiff\Utils.h" />
-    <ClInclude Include="types\include\types\complex.h" />
+    <ClInclude Include="types\include\types\Complex.h" />
     <ClInclude Include="types\include\types\PageRowCol.h" />
     <ClInclude Include="types\include\types\Range.h" />
     <ClInclude Include="types\include\types\RangeList.h" />
