+++ modules/c++/str/unittests/test_str.cpp
@@ -22,7 +22,7 @@
 
 #include <tuple> // std::ignore
 
-#include <types/complex.h>
+#include <types/Complex.h>
 #include <config/compiler_extensions.h>
 #include <import/str.h>
